requirements.txt
flask==3.0.2
requests==2.32.3
pandas==2.2.2
pandas-access==0.0.1
peopledatalabs==5.0.0
jinja2==3.1.4
matplotlib==3.8.4
seaborn==0.13.2
openpyxl==3.1.2
python-dotenv==1.0.0
pathlib2==2.3.7